Move theme code into theme folder.
