+++ crates/ui/src/theme/theme_color.rs
@@ -1,63 +1,9 @@
-use gpui::{
-    hsla, point, px, transparent_black, App, BoxShadow, Global, Hsla, Pixels, SharedString, Window,
-    WindowAppearance,
-};
+use super::color::*;
+use gpui::{transparent_black, Hsla};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::ops::{Deref, DerefMut};
 
-use crate::{colors::*, scroll::ScrollbarShow};
-
-pub fn init(cx: &mut App) {
-    Theme::sync_system_appearance(None, cx);
-    Theme::sync_scrollbar_appearance(cx);
-}
-
-pub trait ActiveTheme {
-    fn theme(&self) -> &Theme;
-}
-
-impl ActiveTheme for App {
-    #[inline(always)]
-    fn theme(&self) -> &Theme {
-        Theme::global(self)
-    }
-}
-
-/// Make a [gpui::Hsla] color.
-///
-/// - h: 0..360.0
-/// - s: 0.0..100.0
-/// - l: 0.0..100.0
-#[inline]
-pub fn hsl(h: f32, s: f32, l: f32) -> Hsla {
-    hsla(h / 360., s / 100.0, l / 100.0, 1.0)
-}
-
-/// Make a BoxShadow like CSS
-///
-/// e.g:
-///
-/// If CSS is `box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);`
-///
-/// Then the equivalent in Rust is `box_shadow(0., 0., 10., 0., hsla(0., 0., 0., 0.1))`
-#[inline]
-pub fn box_shadow(
-    x: impl Into<Pixels>,
-    y: impl Into<Pixels>,
-    blur: impl Into<Pixels>,
-    spread: impl Into<Pixels>,
-    color: Hsla,
-) -> BoxShadow {
-    BoxShadow {
-        offset: point(x.into(), y.into()),
-        blur_radius: blur.into(),
-        spread_radius: spread.into(),
-        color,
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default)]
+#[derive(Debug, Default, Clone, Copy, JsonSchema, Serialize, Deserialize)]
 pub struct ThemeColor {
     /// Used for accents such as hover background on MenuItem, ListItem, etc.
     pub accent: Hsla,
@@ -449,246 +395,6 @@ impl ThemeColor {
             warning_hover: yellow_900().lighten(0.1),
             overlay: white().opacity(0.03),
             window_border: neutral_800(),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Theme {
-    pub colors: ThemeColor,
-
-    pub mode: ThemeMode,
-    pub font_family: SharedString,
-    pub font_size: Pixels,
-    /// Radius for the general elements.
-    pub radius: Pixels,
-    /// Radius for the large elements, e.g.: Modal, Notification border radius.
-    pub radius_lg: Pixels,
-    pub shadow: bool,
-    pub transparent: Hsla,
-    /// Show the scrollbar mode, default: Scrolling
-    pub scrollbar_show: ScrollbarShow,
-    /// Tile grid size, default is 4px.
-    pub tile_grid_size: Pixels,
-    /// The shadow of the tile panel.
-    pub tile_shadow: bool,
-}
-
-impl Deref for Theme {
-    type Target = ThemeColor;
-
-    fn deref(&self) -> &Self::Target {
-        &self.colors
-    }
-}
-
-impl DerefMut for Theme {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.colors
-    }
-}
-
-impl Global for Theme {}
-
-impl Theme {
-    /// Returns the global theme reference
-    #[inline(always)]
-    pub fn global(cx: &App) -> &Theme {
-        cx.global::<Theme>()
-    }
-
-    /// Returns the global theme mutable reference
-    #[inline(always)]
-    pub fn global_mut(cx: &mut App) -> &mut Theme {
-        cx.global_mut::<Theme>()
-    }
-
-    /// Returns true if the theme is dark.
-    #[inline(always)]
-    pub fn is_dark(&self) -> bool {
-        self.mode.is_dark()
-    }
-
-    /// Apply a mask color to the theme.
-    pub fn apply_color(&mut self, mask_color: Hsla) {
-        self.title_bar = self.title_bar.apply(mask_color);
-        self.title_bar_border = self.title_bar_border.apply(mask_color);
-        self.background = self.background.apply(mask_color);
-        self.foreground = self.foreground.apply(mask_color);
-        self.card = self.card.apply(mask_color);
-        self.card_foreground = self.card_foreground.apply(mask_color);
-        self.caret = self.caret.apply(mask_color);
-        self.popover = self.popover.apply(mask_color);
-        self.popover_foreground = self.popover_foreground.apply(mask_color);
-        self.primary = self.primary.apply(mask_color);
-        self.primary_hover = self.primary_hover.apply(mask_color);
-        self.primary_active = self.primary_active.apply(mask_color);
-        self.primary_foreground = self.primary_foreground.apply(mask_color);
-        self.secondary = self.secondary.apply(mask_color);
-        self.secondary_hover = self.secondary_hover.apply(mask_color);
-        self.secondary_active = self.secondary_active.apply(mask_color);
-        self.secondary_foreground = self.secondary_foreground.apply(mask_color);
-        // self.danger = self.danger.apply(mask_color);
-        // self.danger_hover = self.danger_hover.apply(mask_color);
-        // self.danger_active = self.danger_active.apply(mask_color);
-        // self.danger_foreground = self.danger_foreground.apply(mask_color);
-        self.muted = self.muted.apply(mask_color);
-        self.muted_foreground = self.muted_foreground.apply(mask_color);
-        self.accent = self.accent.apply(mask_color);
-        self.accent_foreground = self.accent_foreground.apply(mask_color);
-        self.border = self.border.apply(mask_color);
-        self.input = self.input.apply(mask_color);
-        self.ring = self.ring.apply(mask_color);
-        // self.selection = self.selection.apply(mask_color);
-        self.scrollbar = self.scrollbar.apply(mask_color);
-        self.scrollbar_thumb = self.scrollbar_thumb.apply(mask_color);
-        self.scrollbar_thumb_hover = self.scrollbar_thumb_hover.apply(mask_color);
-        self.drag_border = self.drag_border.apply(mask_color);
-        self.drop_target = self.drop_target.apply(mask_color);
-        self.tab_bar = self.tab_bar.apply(mask_color);
-        self.tab = self.tab.apply(mask_color);
-        self.tab_active = self.tab_active.apply(mask_color);
-        self.tab_foreground = self.tab_foreground.apply(mask_color);
-        self.tab_active_foreground = self.tab_active_foreground.apply(mask_color);
-        self.tab_bar_segmented = self.tab_bar_segmented.apply(mask_color);
-        self.progress_bar = self.progress_bar.apply(mask_color);
-        self.slider_bar = self.slider_bar.apply(mask_color);
-        self.slider_thumb = self.slider_thumb.apply(mask_color);
-        self.list = self.list.apply(mask_color);
-        self.list_even = self.list_even.apply(mask_color);
-        self.list_head = self.list_head.apply(mask_color);
-        self.list_active = self.list_active.apply(mask_color);
-        self.list_active_border = self.list_active_border.apply(mask_color);
-        self.list_hover = self.list_hover.apply(mask_color);
-        self.table = self.table.apply(mask_color);
-        self.table_even = self.table_even.apply(mask_color);
-        self.table_active = self.table_active.apply(mask_color);
-        self.table_active_border = self.table_active_border.apply(mask_color);
-        self.table_hover = self.table_hover.apply(mask_color);
-        self.table_row_border = self.table_row_border.apply(mask_color);
-        self.table_head = self.table_head.apply(mask_color);
-        self.table_head_foreground = self.table_head_foreground.apply(mask_color);
-        self.link = self.link.apply(mask_color);
-        self.link_hover = self.link_hover.apply(mask_color);
-        self.link_active = self.link_active.apply(mask_color);
-        self.skeleton = self.skeleton.apply(mask_color);
-        self.accordion = self.accordion.apply(mask_color);
-        self.accordion_hover = self.accordion_hover.apply(mask_color);
-        self.accordion_active = self.accordion_active.apply(mask_color);
-        self.title_bar = self.title_bar.apply(mask_color);
-        self.title_bar_border = self.title_bar_border.apply(mask_color);
-        self.sidebar = self.sidebar.apply(mask_color);
-        self.sidebar_accent = self.sidebar_accent.apply(mask_color);
-        self.sidebar_accent_foreground = self.sidebar_accent_foreground.apply(mask_color);
-        self.sidebar_border = self.sidebar_border.apply(mask_color);
-        self.sidebar_foreground = self.sidebar_foreground.apply(mask_color);
-        self.sidebar_primary = self.sidebar_primary.apply(mask_color);
-        self.sidebar_primary_foreground = self.sidebar_primary_foreground.apply(mask_color);
-        self.tiles = self.tiles.apply(mask_color);
-        self.description_list_label = self.description_list_label.apply(mask_color);
-        self.description_list_label_foreground =
-            self.description_list_label_foreground.apply(mask_color);
-    }
-
-    /// Sync the theme with the system appearance
-    pub fn sync_system_appearance(window: Option<&mut Window>, cx: &mut App) {
-        // Better use window.appearance() for avoid error on Linux.
-        // https://github.com/longbridge/gpui-component/issues/104
-        let appearance = window
-            .as_ref()
-            .map(|window| window.appearance())
-            .unwrap_or_else(|| cx.window_appearance());
-
-        Self::change(appearance, window, cx);
-    }
-
-    /// Sync the Scrollbar showing behavior with the system
-    pub fn sync_scrollbar_appearance(cx: &mut App) {
-        if cx.should_auto_hide_scrollbars() {
-            cx.global_mut::<Theme>().scrollbar_show = ScrollbarShow::Scrolling;
-        } else {
-            cx.global_mut::<Theme>().scrollbar_show = ScrollbarShow::Hover;
-        }
-    }
-
-    pub fn change(mode: impl Into<ThemeMode>, window: Option<&mut Window>, cx: &mut App) {
-        let mode = mode.into();
-        let colors = match mode {
-            ThemeMode::Light => ThemeColor::light(),
-            ThemeMode::Dark => ThemeColor::dark(),
-        };
-
-        if !cx.has_global::<Theme>() {
-            let theme = Theme::from(colors);
-            cx.set_global(theme);
-        }
-
-        let theme = cx.global_mut::<Theme>();
-
-        theme.mode = mode;
-        theme.colors = colors;
-
-        if let Some(window) = window {
-            window.refresh();
-        }
-    }
-}
-
-impl From<ThemeColor> for Theme {
-    fn from(colors: ThemeColor) -> Self {
-        let mode = ThemeMode::default();
-        Theme {
-            mode,
-            transparent: Hsla::transparent_black(),
-            font_size: px(16.),
-            font_family: if cfg!(target_os = "macos") {
-                ".SystemUIFont".into()
-            } else if cfg!(target_os = "windows") {
-                "Segoe UI".into()
-            } else {
-                "FreeMono".into()
-            },
-            radius: px(6.),
-            radius_lg: px(8.),
-            shadow: true,
-            scrollbar_show: ScrollbarShow::default(),
-            tile_grid_size: px(8.),
-            tile_shadow: true,
-            colors,
-        }
-    }
-}
-
-#[derive(
-    Debug, Clone, Copy, Default, PartialEq, PartialOrd, Eq, Hash, Serialize, Deserialize, JsonSchema,
-)]
-#[serde(rename_all = "snake_case")]
-pub enum ThemeMode {
-    Light,
-    #[default]
-    Dark,
-}
-
-impl ThemeMode {
-    #[inline(always)]
-    pub fn is_dark(&self) -> bool {
-        matches!(self, Self::Dark)
-    }
-
-    /// Return lower_case theme name: `light`, `dark`.
-    pub fn name(&self) -> &'static str {
-        match self {
-            ThemeMode::Light => "light",
-            ThemeMode::Dark => "dark",
-        }
-    }
-}
-
-impl From<WindowAppearance> for ThemeMode {
-    fn from(appearance: WindowAppearance) -> Self {
-        match appearance {
-            WindowAppearance::Dark | WindowAppearance::VibrantDark => Self::Dark,
-            WindowAppearance::Light | WindowAppearance::VibrantLight => Self::Light,
         }
     }
 }
